lab-05/img-hidden/requirements.txt
Pillow
cryptography 
